refactor: rename get_git_root_generic to get_git_root
- Rename module from get_git_root_generic to get_git_root
- Update all imports across handlers, backend, and tests
- Add legacy get_git_root_with_executor function for benchmark compatibility
- Remove the old dynamic dispatch module completely

The generic implementation is now the primary get_git_root module,
simplifying the API and making it clear this is the preferred approach.
