+++ rust/src/git/libs/mod.rs
@@ -6,7 +6,6 @@ pub mod current_commit;
 pub mod get_current_branch;
 pub mod get_current_worktree;
 pub mod get_git_root;
-pub mod get_git_root_generic;
 pub mod is_inside_work_tree;
 pub mod list_branches;
 pub mod list_worktrees;
